WebService Display+Create+Delete

+++ Pidev/Models/SkillModel.cs
@@ -7,8 +7,9 @@ namespace Pidev.Models
 {
     public class SkillModel
     {
-        public long skillId { get; set; }
+        public int skillId { get; set; }
         public string skillDesc { get; set; }
         public string skillName { get; set; }
+        //public DateTime skillDate { get; set; }
     }
 }
